bulkActions checkbox not saving anymore
Since i've updated yiibooster to v.3.1.1 checked boxes does not save anymore while changing page. 
In the $.fn.yiiGridView.initBulkActions function there is no more a "push" inside checkedItems array, so the script isn't saving anything. After changing gridvirew page, all the checked elements are gone, so if i go back to the last gridview page no checkbox is checked.
So, now the select all checkbox is working but checkboxes save between gridview page isn't.
Also illegar javascript character used in value of column is still not checked (maybe there is a better way but at least this is working.
The pop function in checkedItems is wrong, because it causes the deletion of all element after the one unchecked, it's better use slice().